+++ test/core/req_test.dart
@@ -9,7 +9,7 @@ void main() {
     const errorMany = 'always wrong';
     final req = Req(blank: error);
     final reqVal =
-        Req.val(Len.range(1, 3, short: errorVal, long: errorVal), blank: error);
+        Req.val(Len.range(1, 3, less: errorVal, great: errorVal), blank: error);
     final reqStr = Req.str(Digit(non: errorStr), blank: error);
     final reqMany =
         Req.many([const Ok(), const Nok(error: errorMany)], blank: error);